requirements.txt
configparser==4.0.2
elasticsearch==7.0.5
mysql==0.0.2
mysqlclient==1.4.4
urllib3==1.25.6